refactor(LocalImageStorageManager): DI를 통한 의종성 주입 #22
- System.getenv -> application.yaml 환경변수로 변경
