+++ src/main/resources/application-local.yaml
@@ -10,3 +10,6 @@ spring:
       driver-class-name: com.mysql.cj.jdbc.Driver
       username: ${MYSQL_ROOT_USERNAME}
       password: ${MYSQL_ROOT_PASSWORD}
+
+app:
+  image-base-path: ${LOCAL_IMAGE_STORAGE}
